+++ userspace/lib/include/sys/stat.h
@@ -4,6 +4,7 @@
 #include <sys/types.h>
 
 #define S_IFMT 0170000
+#define S_IFIFO 0010000
 #define S_IFDIR 0040000
 #define S_IFREG 0100000
 #define S_IRUSR 0400
@@ -20,6 +21,7 @@
 #define S_IRWXO 0007
 
 #define S_ISDIR(mode) (((mode) & S_IFMT) == S_IFDIR)
+#define S_ISFIFO(mode) (((mode) & S_IFMT) == S_IFIFO)
 #define S_ISREG(mode) (((mode) & S_IFMT) == S_IFREG)
 
 struct stat {